+++ ent/problem/where.go
@@ -65,17 +65,17 @@ func Content(v string) predicate.Problem {
 }
 
 // Point applies equality check predicate on the "point" field. It's identical to PointEQ.
-func Point(v int) predicate.Problem {
+func Point(v int16) predicate.Problem {
 	return predicate.Problem(sql.FieldEQ(FieldPoint, v))
 }
 
 // CaseVersion applies equality check predicate on the "case_version" field. It's identical to CaseVersionEQ.
-func CaseVersion(v int) predicate.Problem {
+func CaseVersion(v int16) predicate.Problem {
 	return predicate.Problem(sql.FieldEQ(FieldCaseVersion, v))
 }
 
 // Index applies equality check predicate on the "index" field. It's identical to IndexEQ.
-func Index(v int) predicate.Problem {
+func Index(v int16) predicate.Problem {
 	return predicate.Problem(sql.FieldEQ(FieldIndex, v))
 }
 
@@ -220,122 +220,122 @@ func ContentContainsFold(v string) predicate.Problem {
 }
 
 // PointEQ applies the EQ predicate on the "point" field.
-func PointEQ(v int) predicate.Problem {
+func PointEQ(v int16) predicate.Problem {
 	return predicate.Problem(sql.FieldEQ(FieldPoint, v))
 }
 
 // PointNEQ applies the NEQ predicate on the "point" field.
-func PointNEQ(v int) predicate.Problem {
+func PointNEQ(v int16) predicate.Problem {
 	return predicate.Problem(sql.FieldNEQ(FieldPoint, v))
 }
 
 // PointIn applies the In predicate on the "point" field.
-func PointIn(vs ...int) predicate.Problem {
+func PointIn(vs ...int16) predicate.Problem {
 	return predicate.Problem(sql.FieldIn(FieldPoint, vs...))
 }
 
 // PointNotIn applies the NotIn predicate on the "point" field.
-func PointNotIn(vs ...int) predicate.Problem {
+func PointNotIn(vs ...int16) predicate.Problem {
 	return predicate.Problem(sql.FieldNotIn(FieldPoint, vs...))
 }
 
 // PointGT applies the GT predicate on the "point" field.
-func PointGT(v int) predicate.Problem {
+func PointGT(v int16) predicate.Problem {
 	return predicate.Problem(sql.FieldGT(FieldPoint, v))
 }
 
 // PointGTE applies the GTE predicate on the "point" field.
-func PointGTE(v int) predicate.Problem {
+func PointGTE(v int16) predicate.Problem {
 	return predicate.Problem(sql.FieldGTE(FieldPoint, v))
 }
 
 // PointLT applies the LT predicate on the "point" field.
-func PointLT(v int) predicate.Problem {
+func PointLT(v int16) predicate.Problem {
 	return predicate.Problem(sql.FieldLT(FieldPoint, v))
 }
 
 // PointLTE applies the LTE predicate on the "point" field.
-func PointLTE(v int) predicate.Problem {
+func PointLTE(v int16) predicate.Problem {
 	return predicate.Problem(sql.FieldLTE(FieldPoint, v))
 }
 
 // CaseVersionEQ applies the EQ predicate on the "case_version" field.
-func CaseVersionEQ(v int) predicate.Problem {
+func CaseVersionEQ(v int16) predicate.Problem {
 	return predicate.Problem(sql.FieldEQ(FieldCaseVersion, v))
 }
 
 // CaseVersionNEQ applies the NEQ predicate on the "case_version" field.
-func CaseVersionNEQ(v int) predicate.Problem {
+func CaseVersionNEQ(v int16) predicate.Problem {
 	return predicate.Problem(sql.FieldNEQ(FieldCaseVersion, v))
 }
 
 // CaseVersionIn applies the In predicate on the "case_version" field.
-func CaseVersionIn(vs ...int) predicate.Problem {
+func CaseVersionIn(vs ...int16) predicate.Problem {
 	return predicate.Problem(sql.FieldIn(FieldCaseVersion, vs...))
 }
 
 // CaseVersionNotIn applies the NotIn predicate on the "case_version" field.
-func CaseVersionNotIn(vs ...int) predicate.Problem {
+func CaseVersionNotIn(vs ...int16) predicate.Problem {
 	return predicate.Problem(sql.FieldNotIn(FieldCaseVersion, vs...))
 }
 
 // CaseVersionGT applies the GT predicate on the "case_version" field.
-func CaseVersionGT(v int) predicate.Problem {
+func CaseVersionGT(v int16) predicate.Problem {
 	return predicate.Problem(sql.FieldGT(FieldCaseVersion, v))
 }
 
 // CaseVersionGTE applies the GTE predicate on the "case_version" field.
-func CaseVersionGTE(v int) predicate.Problem {
+func CaseVersionGTE(v int16) predicate.Problem {
 	return predicate.Problem(sql.FieldGTE(FieldCaseVersion, v))
 }
 
 // CaseVersionLT applies the LT predicate on the "case_version" field.
-func CaseVersionLT(v int) predicate.Problem {
+func CaseVersionLT(v int16) predicate.Problem {
 	return predicate.Problem(sql.FieldLT(FieldCaseVersion, v))
 }
 
 // CaseVersionLTE applies the LTE predicate on the "case_version" field.
-func CaseVersionLTE(v int) predicate.Problem {
+func CaseVersionLTE(v int16) predicate.Problem {
 	return predicate.Problem(sql.FieldLTE(FieldCaseVersion, v))
 }
 
 // IndexEQ applies the EQ predicate on the "index" field.
-func IndexEQ(v int) predicate.Problem {
+func IndexEQ(v int16) predicate.Problem {
 	return predicate.Problem(sql.FieldEQ(FieldIndex, v))
 }
 
 // IndexNEQ applies the NEQ predicate on the "index" field.
-func IndexNEQ(v int) predicate.Problem {
+func IndexNEQ(v int16) predicate.Problem {
 	return predicate.Problem(sql.FieldNEQ(FieldIndex, v))
 }
 
 // IndexIn applies the In predicate on the "index" field.
-func IndexIn(vs ...int) predicate.Problem {
+func IndexIn(vs ...int16) predicate.Problem {
 	return predicate.Problem(sql.FieldIn(FieldIndex, vs...))
 }
 
 // IndexNotIn applies the NotIn predicate on the "index" field.
-func IndexNotIn(vs ...int) predicate.Problem {
+func IndexNotIn(vs ...int16) predicate.Problem {
 	return predicate.Problem(sql.FieldNotIn(FieldIndex, vs...))
 }
 
 // IndexGT applies the GT predicate on the "index" field.
-func IndexGT(v int) predicate.Problem {
+func IndexGT(v int16) predicate.Problem {
 	return predicate.Problem(sql.FieldGT(FieldIndex, v))
 }
 
 // IndexGTE applies the GTE predicate on the "index" field.
-func IndexGTE(v int) predicate.Problem {
+func IndexGTE(v int16) predicate.Problem {
 	return predicate.Problem(sql.FieldGTE(FieldIndex, v))
 }
 
 // IndexLT applies the LT predicate on the "index" field.
-func IndexLT(v int) predicate.Problem {
+func IndexLT(v int16) predicate.Problem {
 	return predicate.Problem(sql.FieldLT(FieldIndex, v))
 }
 
 // IndexLTE applies the LTE predicate on the "index" field.
-func IndexLTE(v int) predicate.Problem {
+func IndexLTE(v int16) predicate.Problem {
 	return predicate.Problem(sql.FieldLTE(FieldIndex, v))
 }
